refactor: split account settings into PasskeySection and TwoFactorSection
- Extract passkey management (load/register/delete + state + UI) to PasskeySection.svelte
- Extract 2FA management (enable/verify/disable + state + UI) to TwoFactorSection.svelte
- account/+page.svelte: 543 → 224 lines (name edit, Google link, layout only)
- twoFactorEnabled/invalidateAll passed via props for same behavior
- Verified: svelte-check 0, eslint 0, tests 61/61
